Post History fix
This should fix optics on new posts. Wasn't copying prior arrays over properly. Also I was accidentally flooding my promises array because I didn't clean it up, which is why I moved it down to the scope it's at now. We now only display applicants when there is an actual change to an applicant status.